refactor : s3 기능 변경에 따른 API 및 로직 수정 (#159)
