+++ src/main/java/com/example/betteriter/fo_domain/user/controller/AuthController.java
@@ -1,5 +1,7 @@
 package com.example.betteriter.fo_domain.user.controller;
 
+import static com.example.betteriter.global.common.code.status.ErrorStatus._METHOD_ARGUMENT_ERROR;
+
 import com.example.betteriter.fo_domain.user.dto.JoinDto;
 import com.example.betteriter.fo_domain.user.dto.LoginDto;
 import com.example.betteriter.fo_domain.user.dto.PasswordResetRequestDto;
@@ -7,22 +9,25 @@ import com.example.betteriter.fo_domain.user.dto.UserServiceTokenResponseDto;
 import com.example.betteriter.fo_domain.user.exception.UserHandler;
 import com.example.betteriter.fo_domain.user.service.AuthService;
 import com.example.betteriter.global.common.response.ResponseDto;
-import com.example.betteriter.infra.EmailAuthenticationDto;
-import com.example.betteriter.infra.EmailDto;
+import com.example.betteriter.infra.email.EmailAuthenticationDto;
+import com.example.betteriter.infra.email.EmailDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-
-import static com.example.betteriter.global.common.code.status.ErrorStatus._METHOD_ARGUMENT_ERROR;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * - 회원 회원가입/로그인 등 인증,인가와 관련된 작업
@@ -34,25 +39,23 @@ import static com.example.betteriter.global.common.code.status.ErrorStatus._METH
 @RequestMapping("/auth")
 @RestController
 public class AuthController {
+
     private final AuthService authService;
 
     /**
-     * 1. 일반 회원가입
-     * - 서비스에서 사용자 인증 정보 관리 컨트롤러
-     * - 이미 회원가입된 사용자는 에러처리
-     * - /auth/join
+     * 1. 일반 회원가입 - 서비스에서 사용자 인증 정보 관리 컨트롤러 - 이미 회원가입된 사용자는 에러처리 - /auth/join
      **/
     @PostMapping("/join")
     @Operation(summary = "일반 회원가입")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "SUCCESS_200", description = "일반 회원가입 성공"),
-            @ApiResponse(responseCode = "AUTH_EMAIL_DUPLICATION_401", description = "이메일 이미 존재"),
-            @ApiResponse(responseCode = "METHOD_ARGUMENT_ERROR", description = "올바르지 않은 클라이언트 요청값")}
+        @ApiResponse(responseCode = "SUCCESS_200", description = "일반 회원가입 성공"),
+        @ApiResponse(responseCode = "AUTH_EMAIL_DUPLICATION_401", description = "이메일 이미 존재"),
+        @ApiResponse(responseCode = "METHOD_ARGUMENT_ERROR", description = "올바르지 않은 클라이언트 요청값")}
     )
     public ResponseDto<Long> join(
-            @Schema(description = "일반 회원가입 요청 객체")
-            @RequestBody @Valid JoinDto request,
-            BindingResult bindingResult
+        @Schema(description = "일반 회원가입 요청 객체")
+        @RequestBody @Valid JoinDto request,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         return ResponseDto.onSuccess(this.authService.join(request));
@@ -60,17 +63,14 @@ public class AuthController {
 
 
     /**
-     * 2. 회원가입을 위한 이메일 인증 요청
-     * - 사용자가 일반 회원가입에서 입력한 이메일을 통한 인증 컨트롤러
-     * - 해당 이메일 중복 여부 판단 후, 인증 번호 전송
-     * 과정 : 클라이언트 이메일 입력 및 인증 버튼 클릭 -> 서버로 해당 이메일과 함께 API 요청 ->
-     * -> 해당 이메일을 가진 회원이 있는지 중복 체크 -> 없다면 인증 코드 이메일로 전송 ->
-     * 사용자가 인증 번호 클릭 -> 서버에서 판단 후 승인 or 거부
+     * 2. 회원가입을 위한 이메일 인증 요청 - 사용자가 일반 회원가입에서 입력한 이메일을 통한 인증 컨트롤러 - 해당 이메일 중복 여부 판단 후, 인증 번호 전송 과정 : 클라이언트 이메일 입력 및 인증
+     * 버튼 클릭 -> 서버로 해당 이메일과 함께 API 요청 -> -> 해당 이메일을 가진 회원이 있는지 중복 체크 -> 없다면 인증 코드 이메일로 전송 -> 사용자가 인증 번호 클릭 -> 서버에서 판단 후
+     * 승인 or 거부
      */
     @PostMapping("/join/emails")
     public ResponseDto<Void> requestEmail(
-            @Valid @RequestBody EmailDto emailDto,
-            BindingResult bindingResult
+        @Valid @RequestBody EmailDto emailDto,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         this.authService.requestEmailForJoin(emailDto);
@@ -78,14 +78,12 @@ public class AuthController {
     }
 
     /**
-     * 3. 회원가입 인증 코드 검증
-     * - 사용자가 이메일로 받은 인증 코드 검증하는 컨트롤러
-     * - /auth/check/code
+     * 3. 회원가입 인증 코드 검증 - 사용자가 이메일로 받은 인증 코드 검증하는 컨트롤러 - /auth/check/code
      **/
     @PostMapping("/join/emails/verification")
     public ResponseDto<Void> verifyJoinAuthCode(
-            @Valid @RequestBody EmailAuthenticationDto emailAuthenticationDto,
-            BindingResult bindingResult
+        @Valid @RequestBody EmailAuthenticationDto emailAuthenticationDto,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         this.authService.verifyJoinAuthCode(emailAuthenticationDto);
@@ -94,27 +92,24 @@ public class AuthController {
 
 
     /**
-     * 4. 일반 로그인
-     * - /auth/login
+     * 4. 일반 로그인 - /auth/login
      **/
     @PostMapping("/login")
     public ResponseDto<UserServiceTokenResponseDto> login(
-            @Valid @RequestBody LoginDto loginRequestDto,
-            BindingResult bindingResult
+        @Valid @RequestBody LoginDto loginRequestDto,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         return ResponseDto.onSuccess(this.authService.login(loginRequestDto));
     }
 
     /**
-     * 5. 비밀번호 재설정 이메일 인증 요청 컨트롤러
-     * - /auth/password/emails
-     * - 비밀번호 재설정을 위한 이메일 요청 이미 해당 회원정보가 있어야 함
+     * 5. 비밀번호 재설정 이메일 인증 요청 컨트롤러 - /auth/password/emails - 비밀번호 재설정을 위한 이메일 요청 이미 해당 회원정보가 있어야 함
      **/
     @PostMapping("/password/emails")
     public ResponseDto<Void> requestEmailForPassword(
-            @Valid @RequestBody EmailDto emailDto,
-            BindingResult bindingResult
+        @Valid @RequestBody EmailDto emailDto,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         this.authService.requestEmailForPasswordReset(emailDto);
@@ -123,13 +118,12 @@ public class AuthController {
 
 
     /**
-     * 비밀번호 재설정 인증 코드 검증
-     * - 사용자가 이메일로 받은 인증 코드 검증하는 컨트롤러
+     * 비밀번호 재설정 인증 코드 검증 - 사용자가 이메일로 받은 인증 코드 검증하는 컨트롤러
      **/
     @PostMapping("/password/emails/verification")
     public ResponseDto<Void> verifyPasswordResetAuthCode(
-            @Valid @RequestBody EmailAuthenticationDto emailAuthenticationDto,
-            BindingResult bindingResult
+        @Valid @RequestBody EmailAuthenticationDto emailAuthenticationDto,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         this.authService.verifyPasswordResetAuthCode(emailAuthenticationDto);
@@ -137,16 +131,13 @@ public class AuthController {
     }
 
     /**
-     * 6. 비밀번호 재설정 요청 컨트롤러
-     * - /auth/password/reset
-     * - 비밀번호 이메일 인증 /password/emails 후 진행
-     * - 새로운 비밀번호를 입력받아 설정
-     * - 해당 유저가 존재 여부 확인 및 일반 회원가입 유저인지 다시 확인
+     * 6. 비밀번호 재설정 요청 컨트롤러 - /auth/password/reset - 비밀번호 이메일 인증 /password/emails 후 진행 - 새로운 비밀번호를 입력받아 설정 - 해당 유저가 존재 여부
+     * 확인 및 일반 회원가입 유저인지 다시 확인
      **/
     @PatchMapping("/password/reset")
     public ResponseDto<Void> resetPassword(
-            @Valid @RequestBody PasswordResetRequestDto request,
-            BindingResult bindingResult
+        @Valid @RequestBody PasswordResetRequestDto request,
+        BindingResult bindingResult
     ) {
         this.checkRequestValidation(bindingResult);
         this.authService.resetPassword(request);
@@ -154,14 +145,11 @@ public class AuthController {
     }
 
     /**
-     * 7. 닉네임 중복 여부 판단 컨트롤러
-     * - /auth/nickname/check
-     * - 닉네임 중복 되지 않는 경우(사용 가능) -> true
-     * - 닉네임 중복 되는 경우(사용 불가능) -> false
+     * 7. 닉네임 중복 여부 판단 컨트롤러 - /auth/nickname/check - 닉네임 중복 되지 않는 경우(사용 가능) -> true - 닉네임 중복 되는 경우(사용 불가능) -> false
      **/
     @GetMapping("/nickname/check")
     public ResponseDto<Boolean> checkNickname(
-            @RequestParam String nickname
+        @RequestParam String nickname
     ) {
         return ResponseDto.onSuccess(this.authService.checkNickname(nickname));
     }
